% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor2curve.R
\name{cor2curve}
\alias{cor2curve}
\title{Interpolate the (non)linear shape of a relation between two traits based on genetic correlations}
\usage{
cor2curve(
  rg,
  b1,
  b2,
  se,
  method = "polynomial",
  q1 = NULL,
  q2 = NULL,
  plot = TRUE,
  boot = FALSE,
  x.trait = "",
  y.trait = ""
)
}
\arguments{
\item{rg}{genetic correlation genetic correlation between outcome y and a pair of bins of x}

\item{b1}{value assigned to bin b1 used in the GWAS of "x" usually a count, or median/mean phenotype for participants in the bin, used to compute dx}

\item{b2}{value assigned to bin b2 used in the GWAS of "x" , usually a count, or median/mean phenotype for participants in the bin, used to compute dx}

\item{method}{either "polynomial' or "cubic-spline" specifies the method used for interpolation}
}
\value{
plot of the interpolated curves (optional)

plot of resampled verison of  interpolated curves to indicate statistical uncertainty (optional)

a list with parameters, optimizer statses and the method used for interpolation
}
\description{
Based on genetic correlations between trait y (can be binary or continuous trait) and a number of GWASs of trait x, infer the shap of their relation

Prior to using thes function x was binned into segments of the distribution of x (say deciles), and each unqiuque pair of bins of x are used as cases and controls respectively in a GWAS.

Based on these GWASs of segments of x, and a GWAS of y, genetic correlations are computed, either in LDScore regression, or `ldsc()` in GenomicSEM, which implements LD score regression in r.

The genetic correlations, and informatino of which segments of x fed into the GWAS, a curve is interplotatedf using either a 5th order polynomial, or cubic-spline interpolation where a 32th order polynomial is fit to 3 segments of the total distribution of x. 
}
