% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor2curve.R
\name{plot.cor2curve}
\alias{plot.cor2curve}
\title{Plot a curve based on cor2curve}
\usage{
\method{plot}{cor2curve}(
  x,
  xlim = NULL,
  ylim = NULL,
  main = NULL,
  sub = NULL,
  xlab = NULL,
  ylab = NULL,
  boot = NULL
)
}
\arguments{
\item{x}{an object of class 'cor2curve' generated with the cor2curve() function.}

\item{xlim}{numeric vector, optional. The x-axis limits for the plot, defaults to sensible values.}

\item{ylim}{numeric vector, optional. The y-axis limits for the plot, defaults to sensible values.}

\item{main}{character string, optional. The main title for the plot.}

\item{sub}{character string, optional. The subtitle for the plot.}

\item{xlab}{character string, optional. The x-axis label for the plot, as a default its inherited from cor2curve object.}

\item{ylab}{character string, optional. The y-axis label for the plot, as a default its inherited from cor2curve object.}

\item{boot}{logical, optional. Whether to include bootstrapped lines in the plot, as a default its inherited from cor2curve object.}
}
\value{
A plot of the curve.
}
\description{
This function plot a curve based on the input parameters.
}
