% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor2curve.R
\name{print.cor2curve}
\alias{print.cor2curve}
\title{Print a cor2curve object}
\usage{
\method{print}{cor2curve}(x)
}
\arguments{
\item{x}{A cor2curve object}

\item{...}{Additional arguments (not used)}

\item{quote}{A logical indicating whether to quote the output (default: FALSE)}
}
\value{
A list containing the summary information
}
\description{
This function prints a summary of a cor2curve object, which is a custom object
containing the results of a cor2curve analysis. The output includes the method
used, messages, number of observations, coefficients or cubic spline parameters,
cuts, convergence information, value, and intercept.
}
\details{
The function prints different sets of information depending on the method used
to fit the cor2curve. If the method is "polynomial", it prints the coefficients
of the polynomial. If the method is "cubic-spline", it prints the parameters
of the cubic spline.
}
\examples{
# Create a cor2curve object (not shown)
print.cor2curve(x)
}
