% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor2curve.R
\name{cor2curve}
\alias{cor2curve}
\title{Interpolate the (non)linear shape of a relation between two traits based on genetic correlations}
\usage{
cor2curve(
  rg,
  b1,
  b2,
  se,
  method = "polynomial",
  q1 = NULL,
  q2 = NULL,
  plot = TRUE,
  boot = FALSE,
  x.trait = "",
  y.trait = ""
)
}
\arguments{
\item{rg}{genetic correlation genetic correlation between outcome y and a pair of bins of x}

\item{b1}{value assigned to bin b1 used in the GWAS of "x" usually a count, or median/mean phenotype for participants in the bin, used to compute dx}

\item{b2}{value assigned to bin b2 used in the GWAS of "x" , usually a count, or median/mean phenotype for participants in the bin, used to compute dx}

\item{method}{either "polynomial' or "cubic-spline" specifies the method used for interpolation}
}
\value{
plot of the interpolated curves (optional)

plot of resampled verison of  interpolated curves to indicate statistical uncertainty (optional)

a list with parameters, optimizer statses and the method used for interpolation
}
\description{
Based on genetic correlations between trait y (can be binary or continuous trait) and a number of GWASs of trait x, where x is binned  into segments of the distributions of x, and apira of bins are used as case.control in a GWAS
}
