% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor2curve.R
\name{cor2curve}
\alias{cor2curve}
\title{Interpolate the (non)linear shape of a relation between two traits based on genetic correlations}
\usage{
cor2curve(
  rg,
  b1,
  b2,
  se,
  method = "polynomial",
  q1 = NULL,
  q2 = NULL,
  boot = FALSE,
  x.trait = "",
  y.trait = ""
)
}
\arguments{
\item{rg}{genetic correlation genetic correlation between outcome y and a pair of bins of x}

\item{b1}{value assigned to bin b1 used in the GWAS of "x" usually a count, or median/mean phenotype for participants in the bin, used to compute dx}

\item{b2}{value assigned to bin b2 used in the GWAS of "x" , usually a count, or median/mean phenotype for participants in the bin, used to compute dx}

\item{method}{either "polynomial' or "cubic-spline" specifies the method used for interpolation}

\item{q1}{value (in terms of the scale of b1/b2) at which to place the second of two knows for the method 'cubic-spline', ignored if method is 'polynomial', automatically initiated in left at the default value NULL}

\item{boot}{logical used to indicate whether the genetic correlations should be re-sampled 200 times (based on point estimate and s.e.) to estimate parameter uncertainty (standard errors) and introduce curves to a plot to visualize uncertainty  (optional)}
}
\value{
a list with any data pruning undertaken, parameters, optimizer status and the method used for interpolation

(optionally) a curve reflection the estimated non-linear genetic relationship.
}
\description{
Based on genetic correlations between trait y (can be binary or continuous trait) and a number of GWASs of trait x, where x is binned  into segments of the distributions of x, and a pair of bins are used as case.control in a GWAS
}
